- Refactor LikeController store method to use authenticated user ID - Update `LikeController` to handle likes for tweets and comments - Improve notification handling for existing likes - Simplify like creation and notification logic - Remove unnecessary debug line - Optimize `destroy` method for deleting likes

+++ app/Http/Controllers/LikeController.php
@@ -13,24 +13,34 @@ class LikeController extends Controller {
   /**
    * Store a newly created resource in storage.
    */
-  public function store (Request $request) {
+  public function store(Request $request) {
+
+    // dd($request->all(), $request->user());
+
+    $user_id = $request->user()->id;
     if ($request->isTweet) {
-      $like = Like::where('user_id', $request->user_id)
-                  ->where('tweet_id', $request->tweet_id)
-                  ->whereNotNull('tweet_id')
-                  ->first();
+      $like = Like::where('user_id', $user_id)
+        ->where('tweet_id', $request->tweet_id)
+        ->whereNotNull('tweet_id')
+        ->first();
+
+      // If like exists delete its notification then delete the like
       if ($like !== null) {
         $deleted = DB::delete('DELETE FROM notifications WHERE id = ?', [$like->notification_id]);
-        if ($deleted === 1) {
-          $like->delete();
-        }
-      } else {
-        $createdLike = Like::create($request->only(['user_id', 'tweet_id']));
+        if ($deleted === 1) $like->delete();
+      }
+
+      // Else create a new like and notify the target user with it
+      else {
+        $createdLike = Like::create([
+          "user_id"  => $user_id,
+          "tweet_id" => $request->tweet_id
+        ]);
 
         $targetUser = $createdLike->tweet->user;
 
         $targetUser->notify(new LikeNotification(
-          message : "{$createdLike->user->full_name} liked your tweet",
+          message: "{$createdLike->user->full_name} liked your tweet",
           tweet_id: $createdLike->tweet_id
         ));
 
@@ -41,39 +51,42 @@ class LikeController extends Controller {
 
     if ($request->isComment) {
 
-      $like = Like::where('user_id', $request->user_id)
-                  ->where('comment_id', $request->comment_id)
-                  ->whereNotNull('comment_id')
-                  ->first();
+      $like = Like::where('user_id', $user_id)
+        ->where('comment_id', $request->comment_id)
+        ->whereNotNull('comment_id')
+        ->first();
+
+      // If like exists delete its notification then delete the like
       if ($like !== null) {
         $deleted = DB::delete('DELETE FROM notifications WHERE id = ?', [$like->notification_id]);
-        if ($deleted === 1) {
-          $like->delete();
-        }
-      } else {
-        $createdLike = Like::create($request->only(['user_id', 'comment_id']));
+        if ($deleted === 1) $like->delete();
+      }
+
+      // Else create a new like and notify the target user with it
+      else {
+        $createdLike = Like::create([
+          "user_id"    => $user_id,
+          "comment_id" => $request->comment_id
+        ]);
 
         $targetUser = $createdLike->comment->user;
 
         $targetUser->notify(new LikeNotification(
-          message : "{$createdLike->user->full_name} liked your comment",
+          message: "{$createdLike->user->full_name} liked your comment",
           tweet_id: $createdLike->comment->tweet_id
         ));
 
         $createdLike->notification_id = $targetUser->notifications[0]->id; // Store notification UUID
         $createdLike->save();
       }
-
     }
   }
 
   /**
    * Remove the specified resource from storage.
    */
-  public function destroy ($id): RedirectResponse {
-    $like = Like::find($id);
-
-    $like->delete();
+  public function destroy($id): RedirectResponse {
+    Like::find($id)->delete();
 
     return redirect()->back();
   }
